crawler/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
selectolax==0.3.21
pymysql==1.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
apscheduler==3.10.4
croniter==3.0.3
tenacity==9.0.0
bleach==6.1.0
python-multipart==0.0.12
